Django/Wagtail/Bootstrap/Howler.js App Next Iteration of DLBRadio Currently Incomplete
This app was built as a server-side technical and learning exercise on Django and Wagtail.

UI/UX on this is/was an afterthought.

The decision to add Single Page Application functionality came much later, and is incredibly evident in the JavaScript files and the mix-and-match grid/flex CSS.

This web app is built using the following:

Django Web Framework
Wagtail CMS
Howler.js
Bootstrap5
Demo Tracks by Scott Buckley used under the Creative Commons 4 license By Scott Buckley
Demo Artwork generated using Midjourney AI
The data serving and restful API to retrieve data from the models is complete.

The JavaScript that fetches the data on the otherhand... well it works, but it is not remotely close to clean

There will be another version of this project built with a specific intent from the beginning for it to be a single page application, and Django/Wagtail will only serve as the server-side back end for the application.